requirements.txt
#streamlit==1.28.1
##pandas==2.1.3
#pandas==2.2.3
##numpy==1.24.3
#numpy>=2.1,<3
plotly==5.17.0
#scipy==1.11.4
google-generativeai>=0.3.0

streamlit>=1.37,<2          
pandas==2.2.3               
numpy>=2.3,<3               
scipy>=1.14.1,<2            


